Time: 168 ms (75.72%), Space: 107.6 MB (54.45%) - LeetHub

+++ 2348-number-of-zero-filled-subarrays/2348-number-of-zero-filled-subarrays.cpp
@@ -2,7 +2,7 @@ class Solution {
 public:
     long long zeroFilledSubarray(vector<int>& nums) {
         long long ans = 0;
-        long long ct = 0;
+        int ct = 0;
         for(int i = 0; i < nums.size(); i++){
             if(nums[i] == 0){
                 ct++;
